+++ src/Rocketeer/Tasks/Deploy.php
@@ -19,6 +19,9 @@ class Deploy extends Task
 			$this->executeTask('Setup');
 		}
 
+		// Start timer
+		$timerStart = microtime(true);
+
 		// Update current release
 		$release = time();
 		$this->releasesManager->updateCurrentRelease($release);
@@ -51,7 +54,7 @@ class Deploy extends Task
 			$this->share($currentRelease.'/'.$file);
 		}
 
-		return $this->command->info('Successfully deployed release '.$release);
+		return $this->command->info('Successfully deployed release '.$release. ' in '.round(microtime(true) - $timerStart, 4). 's');
 	}
 
 }