+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - The `deploy:check` command now checks PHP extensions for the cache/database/session drivers you set
 - Rocketeer now share logs by default between releases
 - Added a `$silent` flag to make a `Task::run` call silent no matter what
+- Rocketeer now displays how long the deploy took
 
 ### 0.4.0 (stable)